backend-fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
stripe==7.8.0
openai==1.3.8
google-generativeai==0.3.2
anthropic==0.7.8
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
websockets==12.0
pydantic-settings==2.1.0